This core code was published on the University of Bristol LACMS website in 2008
The original code from 2008 was written and tested on a number of platforms
but I no longer have either a hardware or software environment to test it
in this form.
It was originally in a superfluous directory which has been removed.
The fundamental code has had little in the way of changes since.
There have been changes brought about by supporting software changes
mostly in python, numpy, gcc etc.
The code is research standard software.
It is neither fully optimised nor fully robust.
It has worked on many systems we have applied it to under the conditions
and parameters we have used.
We have done our best to make it general purpose but expect that if you
 apply it to a new problem you may need to do anything from resetting
 tolerance levels to recoding.