The updated version used for the research work from 2008 to 2015
This is an updated version of the code as tested and run on the University
cluster cream. mostly changed to use numpy with fixes for other library
functions.
With a few small changes this was used for the research work from 2008 to 2015.
This cluster is:
Maths Dept's Rocks 6.0 (Mamba) Cluster "CREAM".
CentOS release 6.2 (Final)
Python 2.6.6, numpy.__version__ '1.4.1', gmpy.version() '1.15',
gcc --version gcc (GCC) 4.4.6 20120305 (Red Hat 4.4.6-4)
This will be tagged 0.6.0